requirements.txt
Django>=3.2.0,<6.0.0
djangorestframework>=3.10.0,<4.0.0
psycopg2-binary>=2.8.0
Pillow>=8.0.0 